refactor: Move coverage mocks into its own module
